backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
google-genai
beautifulsoup4
requests
pytest
httpx
